data-analyst-agent/pyproject.toml
[project]
name = "data-analyst-agent"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.70.0",
    "matplotlib>=3.10.8",
    "pandas>=3.0.2",
    "python-dotenv>=1.2.2",
]
